+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CoreMetaData.java
@@ -55,9 +55,14 @@ public class CoreMetaData extends CleverTapMetaData {
 
     private boolean offline;
 
+    /**
+     * Last notification received on device from CleverTap in an active session of a process(before process is killed)
+     */
+    private String lastNotificationId;
+
     private final Object optOutFlagLock = new Object();
 
-    private HashMap<String,Integer> customSdkVersions = new HashMap<>();
+    private HashMap<String, Integer> customSdkVersions = new HashMap<>();
 
     private long referrerClickTime = 0;
 
@@ -69,6 +74,14 @@ public class CoreMetaData extends CleverTapMetaData {
 
     public static Activity getCurrentActivity() {
         return (currentActivity == null) ? null : currentActivity.get();
+    }
+
+    public String getLastNotificationId() {
+        return lastNotificationId;
+    }
+
+    void setLastNotificationId(final String lastNotificationId) {
+        this.lastNotificationId = lastNotificationId;
     }
 
     static int getInitialAppEnteredForegroundTime() {